Rules added, Parsing added

+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: alde-abr <alde-abr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/07 13:16:01 by alexandre         #+#    #+#             */
-/*   Updated: 2025/02/07 11:31:10 by alde-abr         ###   ########.fr       */
+/*   Updated: 2025/02/09 01:36:44 by alde-abr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ typedef unsigned short	t_uint16;
 typedef unsigned int	t_uint32;
 typedef unsigned long	t_uint64;
 
-# define INT_MIN -0x80000000
+# define INT_MIN (int)-0x80000000
 # define INT_MAX 2147483647
 
 //Part 1 : libc Functions
